refactor: move authentication-related tests to a dedicated auth directory
