+++ packages/app/__tests__/controller/auth/signup.js
@@ -1,18 +1,18 @@
 const request = require("supertest");
 const { STATUS_CODES } = require("http");
-const { UserToken } = require("../../models");
+const { UserToken } = require("../../../models");
 const {
   UserTokenService,
   UserService,
   SubscriptionService,
-} = require("../../services");
-const { SIGNUP_PAYLOAD, ENDPOINTS } = require("../../utils/testconstants");
+} = require("../../../services");
+const { SIGNUP_PAYLOAD, ENDPOINTS } = require("../../../utils/testconstants");
 const {
   MOCK_USERS,
   MOCK_SUBSCRIPTION,
   MOCK_USERTOKENS,
-} = require("../../utils/mocks");
-const app = require("../../server");
+} = require("../../../utils/mocks");
+const app = require("../../../server");
 
 describe("SIGNUP API", () => {
   beforeAll(() => {
